requirements.txt
ExifRead==3.0.0
matplotlib==3.5.3
mutagen==1.45.1
pandas==1.4.3
Pillow==9.2.0
PyPDF2==2.10.3
seaborn==0.11.2
streamlit==1.12.0
